+++ jbpm/process-serialization-protobuf/src/main/java/org/jbpm/marshalling/impl/ProtobufProcessMarshaller.java
@@ -35,7 +35,6 @@ import org.drools.core.marshalling.impl.ProcessMarshaller;
 import org.drools.core.process.instance.KogitoWorkItem;
 import org.drools.core.process.instance.KogitoWorkItemManager;
 import org.drools.core.process.instance.impl.KogitoWorkItemImpl;
-import org.drools.serialization.protobuf.PersisterHelper;
 import org.drools.serialization.protobuf.ProtobufMessages;
 import org.drools.serialization.protobuf.ProtobufMessages.Header;
 import org.jbpm.marshalling.impl.JBPMMessages.Variable;
